Move utility functions to sample from distributions
Moved from cocoscore/ml/fasttext_helpers.py and to cocoscore/ml/tools.py
and updated references.
